+++ CHANGELOG.md
@@ -5,17 +5,30 @@ All notable changes to this project will be documented in this file.
 ### Added
 - **Implemented all Permission Endpoints (`/rest/api/latest/`)**:
 
-    - `/rest/api/latest/mypermissions`
-    - `/rest/api/latest/permissions`
+    - `/rest/api/latest/mypermissions` (GET)
+    - `/rest/api/latest/permissions` (GET)
 
 ---
 
 - **Implemented all Application Properties Endpoints (`/rest/api/latest/application-properties`)**:
 
-    - `/rest/api/latest/application-properties`
-    - `/rest/api/latest/application-properties/{id}`
-    - `/rest/api/latest/application-properties/advanced-settings`
+    - `/rest/api/latest/application-properties`(GET)
+    - `/rest/api/latest/application-properties/{id}` (PUT)
+    - `/rest/api/latest/application-properties/advanced-settings` (GET)
 
 ---
+
+- **Implemented all Application Roles Endpoints (`/rest/api/latest/applicationrole`)**:
+
+    - `/rest/api/latest/applicationrole` (GET, PUT)
+    - `/rest/api/latest/applicationrole/{key}` (GET, PUT)
+
+---
+
+- **Implemented all Attachment Endpoints (`/rest/api/latest/attachment`)**:
+
+    - `/rest/api/latest/attachment` 
+    - `/rest/api/latest/attachment/{id}` (GET, DELETE)
+    - `/rest/api/latest/attachment/meta` (GET)